requirements.txt
asgiref==3.3.1
Django==3.1.7
django-address==0.2.5
django-countries==7.1
django-crispy-forms==1.11.1
django-multiforloop==0.2.1
django-phone-field==1.8.1
Pillow==8.1.0
pytz==2021.1
sqlparse==0.4.1
